Initial commit; Made corrections to packets structures to make the initial functionality work: - character creation; - entrance to the world; - Spavn Npcs; - Doodads; - there is no required file EncryptionManager.cs, CSOffsets.cs and SCOffsets.cs for version 5.0.7.0.

+++ AAEmu.Login/Core/Packets/L2C/ACJoinResponsePacket.cs
@@ -5,14 +5,14 @@ namespace AAEmu.Login.Core.Packets.L2C
 {
     public class ACJoinResponsePacket : LoginPacket
     {
-        private readonly byte _reason;
+        private readonly ushort _reason; // in 3.0 byte, in 5.0 ushort
         private readonly byte _authId;
         private readonly uint _afs;
         private readonly short _unk2;
         private readonly byte _unk3;
         private readonly byte _slotCount;
 
-        public ACJoinResponsePacket(byte authId, byte reason, uint afs, byte slotCount) : base(LCOffsets.ACJoinResponsePacket)
+        public ACJoinResponsePacket(byte authId, ushort reason, uint afs, byte slotCount) : base(LCOffsets.ACJoinResponsePacket)
         {
             _reason = reason;
             _afs = afs;
